Title: Adding a Cron Build Trigger column to Extra columns plugin
Description: This change adds the ability to add a new column to jenkins dashboards,
showing the periodic build trigger for jobs, if it exists. The column will show
the periodic build trigger in crontab format.

Most of the code is adapted from the build description column.

Test Plan: Tested the following cases with a local debug run of jenkins using maven.

1/ Verify that a job with no periodic build trigger shows up empty.
2/ Verify that a job with periodic build trigger shows up correctly.
3/ Verify that removing and adding a trigger has the right effect on the column.